requirements.txt
aniso8601==9.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
gevent==21.1.2
greenlet==1.0.0
gunicorn==19.9.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
jwt==1.2.0
MarkupSafe==1.1.1
pipenv==2020.11.15
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.3
pyrsistent==0.17.3
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
typing-extensions==3.10.0.0
urllib3==1.26.3
virtualenv-clone==0.5.4
Werkzeug==0.16.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.2.0
